test(blocking): apply JA4_r structural matcher to test_tlsfingerprint_blocking too
The previous commit (7ad1723) only patched tests/test_ja4_fingerprints.py.
The blocking test module (tests/test_tlsfingerprint_blocking.py) still
contained 5 exact-equality JA4_r assertions, including
test_ja4r_tls12_fingerprint_exact_match which was the actual CI failure
(production tls.peet.ws emits 't12d128h2' for 12 ciphers + 8 extensions
unpadded; the local tlsfingerprint.com Docker emits the spec-padded
't12d1208h2'; equality fails).

Promote parse_ja4r and assert_ja4r_equivalent helpers from
test_ja4_fingerprints.py into tests/conftest.py as public names so they
can be reused. Both test files now import from conftest. Replace the
five JA4_r equality assertions in test_tlsfingerprint_blocking.py with
structural matcher calls.
