+++ tests/test_ja4_fingerprints.py
@@ -7,141 +7,25 @@ custom JA4_r parameters, and comparison with JA3 fingerprints.
 Based on: /Users/dannydasilva/Documents/personal/CycleTLS/tests/ja4-fingerprint.test.js
 """
 import os
-import re
 
 import pytest
+
+# Structural JA4_r matchers live in tests/conftest.py so they can be reused by
+# test_tlsfingerprint_blocking.py. See conftest for full rationale on why we
+# match structure rather than exact strings (production tls.peet.ws strips
+# leading zeros in the cipher_count/ext_count header field).
+from conftest import (
+    assert_ja4r_equivalent as _assert_ja4r_equivalent,
+)
+from conftest import (
+    parse_ja4r as _parse_ja4r,
+)
 
 from cycletls import CycleTLS
 
 _TLSFP_URL = os.environ.get("TLSFP_URL", "https://tls.peet.ws")
 
 pytestmark = pytest.mark.live
-
-
-# JA4_r header format: t<TLS_ver>d<cipher_count><ext_count><ALPN>
-# Per the JA4 spec, cipher_count and ext_count are 2-digit zero-padded.
-# Production tls.peet.ws emits an unpadded form (e.g. "t12d128h2" for 12+8),
-# while local tlsfingerprint.com Docker emits the spec form ("t12d1208h2").
-# Both are accepted: tests validate STRUCTURE rather than exact prefixes.
-_JA4R_HEADER_RE = re.compile(r"^t(?P<ver>\d{2})d(?P<counts>\d+)(?P<alpn>h2|h1|http)$")
-
-
-def _parse_ja4r(s: str) -> dict:
-    """
-    Parse a JA4_r string into its structural components.
-
-    JA4_r format: t<ver>d<cipher_count><ext_count><ALPN>_<ciphers>_<extensions>_<sigalgs>
-
-    The cipher_count and ext_count fields in the header may be either:
-      - Unpadded (e.g. "128" -- 12 ciphers + 8 extensions, the format
-        currently produced by the production tls.peet.ws server)
-      - Zero-padded to 2 digits each (e.g. "1208" -- 12 + 08, per the JA4
-        spec, the format produced by the local tlsfingerprint.com Docker
-        server)
-
-    Note: the cipher_count and ext_count *header* fields refer to the
-    counts seen on the wire and may include SNI (0x0000) and ALPN (0x0010),
-    while the rendered extension list excludes those. So header counts will
-    NOT always equal `len(extensions)`. This helper returns the header
-    counts as ints (best-effort interpretation, preferring the spec
-    zero-padded form when ambiguous) and the observed list lengths
-    separately.
-
-    Returns a dict with keys:
-      tls_version, alpn, header_cipher_count, header_ext_count,
-      ciphers, extensions, sig_algs, header, raw.
-    """
-    parts = s.split("_")
-    assert len(parts) == 4, f"JA4_r should have 4 underscore-separated parts, got {len(parts)}: {s}"
-
-    header, ciphers_s, exts_s, sigs_s = parts
-    m = _JA4R_HEADER_RE.match(header)
-    assert m, f"JA4_r header malformed: {header!r}"
-
-    ciphers = [c for c in ciphers_s.split(",") if c]
-    extensions = [e for e in exts_s.split(",") if e]
-    sig_algs = [a for a in sigs_s.split(",") if a]
-
-    counts = m.group("counts")
-    # Decode the counts field. Spec form is 2-digit padded each (4 chars).
-    # Production tls.peet.ws strips leading zeros, so a 3-char "128" can
-    # mean 12 ciphers + 8 extensions OR 1 cipher + 28 extensions. We
-    # disambiguate by preferring the interpretation whose cipher count
-    # matches the observed cipher list length (which is invariant across
-    # servers).
-    header_cc, header_ec = _decode_counts(counts, len(ciphers))
-
-    return {
-        "tls_version": m.group("ver"),
-        "alpn": m.group("alpn"),
-        "header_cipher_count": header_cc,
-        "header_ext_count": header_ec,
-        "ciphers": ciphers,
-        "extensions": extensions,
-        "sig_algs": sig_algs,
-        "header": header,
-        "raw": s,
-    }
-
-
-def _decode_counts(counts: str, observed_cipher_count: int) -> tuple[int, int]:
-    """
-    Decode the concatenated cipher_count + ext_count field from a JA4_r
-    header. Returns (cipher_count, ext_count).
-
-    Strategy: enumerate every (cc, ec) split where cc is a prefix of
-    `counts`, prefer the split where cc equals the observed cipher count
-    (this disambiguates unpadded production output). Otherwise fall back
-    to the spec form (2-digit padded each, length 4).
-    """
-    candidates: list[tuple[int, int]] = []
-    for split in range(1, len(counts)):
-        try:
-            cc = int(counts[:split])
-            ec = int(counts[split:])
-        except ValueError:
-            continue
-        candidates.append((cc, ec))
-
-    # Prefer the candidate whose cipher count matches what we actually saw.
-    for cc, ec in candidates:
-        if cc == observed_cipher_count:
-            return cc, ec
-
-    # Fall back to the spec form (4-char zero-padded) if available.
-    if len(counts) == 4:
-        return int(counts[:2]), int(counts[2:])
-
-    # Last resort: assume single-digit cipher count.
-    if candidates:
-        return candidates[0]
-    raise AssertionError(f"Could not decode JA4_r counts field: {counts!r}")
-
-
-def _assert_ja4r_equivalent(actual: str, expected: str) -> None:
-    """
-    Assert two JA4_r strings are structurally equivalent.
-
-    Header padding for cipher_count/ext_count may differ between servers
-    (production unpadded vs spec-compliant zero-padded), but the body
-    (ciphers, extensions, signature algorithms) and TLS version + ALPN
-    must match exactly.
-    """
-    a = _parse_ja4r(actual)
-    e = _parse_ja4r(expected)
-    assert a["tls_version"] == e["tls_version"], (
-        f"TLS version mismatch: actual={a['tls_version']} expected={e['tls_version']}"
-    )
-    assert a["alpn"] == e["alpn"], f"ALPN mismatch: actual={a['alpn']} expected={e['alpn']}"
-    assert a["ciphers"] == e["ciphers"], (
-        f"Cipher list mismatch:\nactual:   {a['ciphers']}\nexpected: {e['ciphers']}"
-    )
-    assert a["extensions"] == e["extensions"], (
-        f"Extension list mismatch:\nactual:   {a['extensions']}\nexpected: {e['extensions']}"
-    )
-    assert a["sig_algs"] == e["sig_algs"], (
-        f"Signature algorithm list mismatch:\nactual:   {a['sig_algs']}\nexpected: {e['sig_algs']}"
-    )
 
 
 @pytest.fixture(scope="module")
